Refactor integration tests into multiple files
