--- tests/README.md
@@ -1,5 +1,0 @@
-# Tests
-
-This directory contains integration tests for Weld.
-
-`cargo test` runs all integration tests.
